stubbenshopper/requirements.txt
pymongo[srv]
flask_bcrypt
flask_jwt_extended
flask_mongoengine
flask_restplus
requests